+++ src/atlas/views/center/results.ts
@@ -1,0 +1,7 @@
+import { EditorView } from '../editor_view';
+
+export class Results extends EditorView {
+	constructor() {
+		super('atlas-engine.results', 'Results');
+	}
+}