+++ src/atlas/views/center/scene.ts
@@ -1,0 +1,7 @@
+import { EditorView } from '../editor_view';
+
+export class Scene extends EditorView {
+	constructor() {
+		super('atlas-engine.scene', 'Scene');
+	}
+}