+++ src/atlas/views/layout.ts
@@ -1,0 +1,58 @@
+import type * as vscode from 'vscode';
+import type { View } from './view';
+import type { EditorView } from './editor_view';
+import type { ViewContainer } from './view_container';
+import { Left } from './left/left';
+import { Right } from './right/right';
+import { Bottom } from './bottom/bottom';
+import { Center } from './center/center';
+
+export class Layout implements vscode.Disposable {
+	private disposed = false;
+	readonly left = new Left();
+	readonly right = new Right();
+	readonly bottom = new Bottom();
+	readonly center = new Center();
+
+	get containers(): readonly ViewContainer[] {
+		return [this.left, this.right, ...this.bottom.containers];
+	}
+
+	get views(): readonly (View | EditorView)[] {
+		return [...this.containers.flatMap(container => container.views), ...this.center.views];
+	}
+
+	initialize(api: typeof vscode): void {
+		for (const view of this.views) {
+			view.initialize(api);
+		}
+	}
+
+	update(): void {
+		for (const view of this.views) {
+			view.update();
+		}
+	}
+
+	async show(api: typeof vscode): Promise<void> {
+		for (const container of [this.left, this.right, this.bottom.containers[0]]) {
+			if (this.disposed) {
+				return;
+			}
+			await api.commands.executeCommand(`${container.views[0].id}.open`, { preserveFocus: true });
+		}
+		for (const view of [...this.center.views].reverse()) {
+			if (this.disposed) {
+				return;
+			}
+			await view.show(api);
+		}
+	}
+
+	dispose(): void {
+		this.disposed = true;
+		for (const view of [...this.views].reverse()) {
+			view.dispose();
+		}
+	}
+}