+++ src/atlas/views/left/assets.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Assets extends View {
+	constructor() {
+		super('atlas-engine.assets', 'ASSETS');
+	}
+}