+++ src/atlas/views/left/boundaries.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Boundaries extends View {
+	constructor() {
+		super('atlas-engine.boundaries', 'BOUNDARIES');
+	}
+}