+++ src/atlas/views/left/domain.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Domain extends View {
+	constructor() {
+		super('atlas-engine.domain', 'DOMAIN');
+	}
+}