+++ src/atlas/views/left/geometry.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Geometry extends View {
+	constructor() {
+		super('atlas-engine.geometry', 'GEOMETRY');
+	}
+}