+++ src/atlas/views/left/materials.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Materials extends View {
+	constructor() {
+		super('atlas-engine.materials', 'MATERIALS');
+	}
+}