+++ src/atlas/views/left/sinks.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Sinks extends View {
+	constructor() {
+		super('atlas-engine.sinks', 'SINKS');
+	}
+}