+++ src/atlas/views/left/sources.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Sources extends View {
+	constructor() {
+		super('atlas-engine.sources', 'SOURCES');
+	}
+}