+++ src/atlas/views/right/inspector.ts
@@ -1,0 +1,7 @@
+import { View } from '../view';
+
+export class Inspector extends View {
+	constructor() {
+		super('atlas-engine.inspector', 'INSPECTOR');
+	}
+}